// src/interfaces/interfaces.ts
export interface CreateClientDTO {
    name: string;
    email: string;
    age: number;
    cpf: string;
    phone: string;
    serasaPFScore: number;
    companyName: string;
    businessArea: string;
    cnpj: string;
    foundationYears: number;
    socialCapital: number;
    serasaCompanyScore: number;
    annualRevenue: number;
    criminalModel: boolean;
    debtHistory: boolean;
  }
  